Tuning after using it at work
- SYMBOL Swapped ? and !
- NUMBER replace backspace with space
- NUMBER added grave back
- NAV redid the right side
- NUMBERS and NAV are now LT, this gives me Bksp and 0 on all layers
- cleaned up unnecessary oled include

Tap Hold
- Optimized the code
- fixed bug in the codes when this feature is off

Combo
- added more delay for CONFIG. Noticed when typing "hope"

RGB
- Made the reset key turn red on CONFIG

Rules
- Oneshot, taphold, and combos are enabled by default

Corne
- Adjusted defaults rules

Cnano
- Adjusted default rules
