+++ keyboards/bastardkb/charybdis/3x5/keymaps/T4CORUN/rules.mk
@@ -6,8 +6,5 @@ USER_NAME := t4corun
 
 #override userspace
 GAMELAYER_ENABLE = yes
-COMBO_ENABLE = yes
-ONESHOT_ENABLE = yes
-TAPHOLD_ENABLE = yes
 
 #override $qmk_firmware\keyboards\bastardkb\charybdis\3x5\v2\splinky_3\rules.mk
