+++ keyboards/crkbd/keymaps/T4CORUN/rules.mk
@@ -7,9 +7,6 @@ OLED_DRIVER = SSD1306
 #override userspace
 # $Repo\qmk_firmware\users\T4CORUN\rules.mk
 DYNAMIC_MACRO_ENABLE = no
-ONESHOT_ENABLE = yes
-TAPHOLD_ENABLE = yes
-COMBO_ENABLE = yes
 
 #override 
 # $qmk_firmware\keyboards\crkbd\rev1\rules.mk
